refactor(datasource): rename datasource_http
Rename the datasource_http module to datasource_ncei to reflect the
source of the data. This is a breaking change for users of the
datasource_http module.
