+++ src/goes_core/datasource/datasource_ncei.py
